+++ ios/RELEASENOTES.md
@@ -1,4 +1,10 @@
 # Release Notes
+### 3.0.0 (21.05.2024)
+* improved load time of the Storyly Widget
+* removed Storyly Moments
+* IMPORTANT! removed StoryMedia class and media field from public Story class
+* IMPORTANT! added previewUrl, actionUrl and storyComponentList to the public Story class
+  
 ### 2.18.1 (20.05.2024)
 * added key field for product variants to indicate variant type